LayoutTests/editing/smart-lists/ordered-markers-separated-by-blank-line-do-not-trigger-conversion.html

<!DOCTYPE html><!-- webkit-test-runner [ SmartListsAvailable=true ] -->
<html>
<head>
<script src="../editing.js"></script>
<script src="../../resources/dump-as-markup.js"></script>
</head>
<body>
<div id="test" contenteditable></div>
<script>
Markup.description('Two ordered-list markers separated by a blank line should not trigger smart-list conversion: the markers must appear on consecutive lines.');

var test = document.getElementById('test');
test.focus();
[...'1. A\n\n2. B'].forEach(typeCharacterCommand);
Markup.dump('test');
</script>
</body>
</html>
