LayoutTests/editing/smart-lists/ordered-markers-separated-by-blank-line-do-not-trigger-conversion-expected.txt

Two ordered-list markers separated by a blank line should not trigger smart-list conversion: the markers must appear on consecutive lines.
| "1. A"
| <div>
|   <br>
| <div>
|   "2. B<#selection-caret>"
